refactor(lru-cache): Convert Node.js tests to TypeScript
