--- packages/lru-cache/.nycrc
@@ -1,19 +1,0 @@
-{
-  "all": false,
-  "branches": 100,
-  "check-coverage": true,
-  "exclude": [
-    "**/*.d.ts",
-    "test/**/*"
-  ],
-  "extension": [
-    ".ts"
-  ],
-  "functions": 100,
-  "lines": 100,
-  "per-file": false,
-  "reporter": [
-    "html"
-  ],
-  "statements": 100
-}
