+++ packages/lru-cache/README.md
@@ -22,4 +22,4 @@ yarn add @wireapp/lru-cache
 
 **Node.js**
 
-- [LRUCacheSpec.js](./test/specs/LRUCacheSpec.js)
+- [LRUCacheSpec.js](src/main/LRUCacheSpec.js)
