+++ packages/lru-cache/tsconfig.json
@@ -1,5 +1,6 @@
 {
   "compilerOptions": {
+    "downlevelIteration": true,
     "outDir": "dist/commonjs",
     "rootDir": "src/main"
   },
